Revert orderTotal / calculateTotal Move inventoryUnit.adjustments to Core model Some minor CS issues
