+++ src/Sylius/Bundle/CartBundle/EventListener/RefreshCartListener.php
@@ -19,10 +19,11 @@ use Symfony\Component\EventDispatcher\GenericEvent;
  */
 class RefreshCartListener
 {
+    /**
+     * @param GenericEvent $event
+     */
     public function refreshCart(GenericEvent $event)
     {
-        // @TODO disable this listener?
-
         $cart = $event->getSubject();
 
         if (!$cart instanceof CartInterface) {
@@ -30,5 +31,7 @@ class RefreshCartListener
                 'RefreshCartListener requires event subject to be instance of "Sylius\Component\Cart\Model\CartInterface"'
             );
         }
+
+        $cart->calculateTotal();
     }
 }
