+++ src/Sylius/Bundle/CoreBundle/spec/EventListener/OrderTaxationListenerSpec.php
@@ -48,6 +48,7 @@ class OrderTaxationListenerSpec extends ObjectBehavior
     ) {
         $event->getSubject()->willReturn($order);
         $taxationProcessor->applyTaxes($order)->shouldBeCalled();
+        $order->calculateTotal()->shouldBeCalled();
 
         $this->applyTaxes($event);
     }
