+++ src/Sylius/Bundle/FixturesBundle/DataFixtures/ORM/LoadOrdersData.php
@@ -66,6 +66,8 @@ class LoadOrdersData extends DataFixture
             $order->setCreatedAt($this->faker->dateTimeBetween('1 year ago', 'now'));
 
             $this->dispatchEvents($order);
+
+            $order->calculateTotal();
             $order->complete();
 
             $paymentState = PaymentInterface::STATE_COMPLETED;
