+++ src/Sylius/Component/Core/Model/InventoryUnit.php
@@ -11,9 +11,12 @@
 
 namespace Sylius\Component\Core\Model;
 
+use Doctrine\Common\Collections\ArrayCollection;
 use Sylius\Component\Inventory\Model\InventoryUnit as BaseInventoryUnit;
+use Sylius\Component\Order\Model\AdjustmentInterface;
 use Sylius\Component\Shipping\Model\ShipmentInterface as BaseShipmentInterface;
 use Sylius\Component\Shipping\Model\ShippableInterface;
+
 
 /**
  * Custom inventory unit class.
@@ -43,6 +46,23 @@ class InventoryUnit extends BaseInventoryUnit implements InventoryUnitInterface
      * @var string ShipmentInterface::STATE_*
      */
     protected $shippingState = ShipmentInterface::STATE_CHECKOUT;
+
+    /**
+     * @var AdjustmentInterface[]
+     */
+    protected $adjustments;
+
+    /**
+     * @var integer
+     */
+    protected $adjustmentsTotal;
+
+    public function __construct()
+    {
+        parent::__construct();
+
+        $this->adjustments = new ArrayCollection();
+    }
 
     /**
      * {@inheritdoc}
@@ -114,5 +134,81 @@ class InventoryUnit extends BaseInventoryUnit implements InventoryUnitInterface
         $this->shippingState = $state;
 
         return $this;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function getAdjustments($type = null)
+    {
+        if (null == $type) {
+            return $this->adjustments;
+        }
+
+        return $this->adjustments->filter(function (AdjustmentInterface $adjustment) use ($type) {
+            return $type === $adjustment->getType();
+        });
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function addAdjustment(AdjustmentInterface $adjustment)
+    {
+        if ($this->adjustments->contains($adjustment)) {
+            return;
+        }
+
+        $adjustment->setAdjustable($this);
+        $this->adjustments->add($adjustment);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function removeAdjustment(AdjustmentInterface $adjustment)
+    {
+        if (!$this->adjustments->contains($adjustment)) {
+            return;
+        }
+
+        $this->adjustments->removeElement($adjustment);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function getAdjustmentsTotal($type = null)
+    {
+        $amount = 0;
+
+        foreach ($this->adjustments as $adjustment) {
+            if ($type && $type !== $adjustment->getType()) {
+                continue;
+            }
+
+            $amount += $adjustment->getAmount();
+        }
+
+        $this->adjustmentsTotal = $amount;
+
+        return $this->adjustmentsTotal;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function removeAdjustments($type)
+    {
+        foreach ($this->getAdjustments($type) as $adjustment) {
+            if ($type === $adjustment->getType() && !$adjustment->isLocked()) {
+                $this->removeAdjustment($adjustment);
+            }
+        }
+    }
+
+    public function clearAdjustments()
+    {
+        $this->adjustments->clear();
     }
 }
