+++ src/Sylius/Component/Core/spec/Model/InventoryUnitSpec.php
@@ -13,6 +13,7 @@ namespace spec\Sylius\Component\Core\Model;
 
 use PhpSpec\ObjectBehavior;
 use Sylius\Component\Core\Model\OrderItemInterface;
+use Sylius\Component\Order\Model\AdjustableInterface;
 use Sylius\Component\Shipping\Model\ShipmentInterface;
 
 class InventoryUnitSpec extends ObjectBehavior
@@ -25,6 +26,11 @@ class InventoryUnitSpec extends ObjectBehavior
     function it_implements_Sylius_core_inventory_unit_interface()
     {
         $this->shouldImplement('Sylius\Component\Core\Model\InventoryUnitInterface');
+    }
+
+    function it_is_adjustable()
+    {
+        $this->shouldImplement(AdjustableInterface::class);
     }
 
     function it_extends_Sylius_inventory_unit_model()
@@ -86,5 +92,142 @@ class InventoryUnitSpec extends ObjectBehavior
 
         $this->setOrderItem(null);
         $this->getOrderItem()->shouldReturn(null);
+    }
+
+    function it_has_no_adjustments_by_default()
+    {
+        $this->getAdjustments()->shouldBeAnInstanceOf('Doctrine\Common\Collections\Collection');
+        $this->getAdjustments()->shouldBeEmpty();
+    }
+
+    function it_allows_to_add_adjustments(
+        AdjustmentInterface $adjustment
+    ) {
+        $this->addAdjustment($adjustment);
+        $this->getAdjustments()->shouldHaveCount(1);
+    }
+
+    function it_allows_to_retrieve_adjustments_by_type(
+        AdjustmentInterface $adjustment,
+        AdjustmentInterface $differentTypeAdjustment
+    ) {
+        $adjustment->getType()->willreturn('type');
+        $differentTypeAdjustment->getType()->willreturn('different_type');
+
+        $adjustment->setAdjustable($this)->shouldBeCalled();
+        $differentTypeAdjustment->setAdjustable($this)->shouldBeCalled();
+
+        $this->addAdjustment($adjustment);
+        $this->addAdjustment($differentTypeAdjustment);
+
+        $this->getAdjustments('type')->shouldHaveCount(1);
+        $this->getAdjustments('different_type')->shouldHaveCount(1);
+    }
+
+    function it_does_not_allow_to_add_same_adjustments(
+        AdjustmentInterface $adjustment
+    ) {
+        $adjustment->setAdjustable($this)->shouldBeCalled();
+
+        $this->getAdjustments()->shouldHaveCount(0);
+        $this->addAdjustment($adjustment);
+        $this->getAdjustments()->shouldHaveCount(1);
+        $this->addAdjustment($adjustment);
+        $this->getAdjustments()->shouldHaveCount(1);
+    }
+
+    function it_allows_removing_adjustments(
+        AdjustmentInterface $adjustment
+    ) {
+        $this->addAdjustment($adjustment);
+        $this->getAdjustments()->shouldHaveCount(1);
+        $this->removeAdjustment($adjustment);
+        $this->getAdjustments()->shouldHaveCount(0);
+    }
+
+    function it_allows_to_know_amount_of_all_adjustments(
+        AdjustmentInterface $adjustment,
+        AdjustmentInterface $adjustment2
+    ) {
+        $adjustment->getAmount()->willReturn(100);
+        $adjustment2->getAmount()->willReturn(300);
+
+        $adjustment->setAdjustable($this)->shouldBeCalled();
+        $adjustment2->setAdjustable($this)->shouldBeCalled();
+
+        $this->addAdjustment($adjustment);
+        $this->addAdjustment($adjustment2);
+
+        $adjustment->setAdjustable($this)->shouldBeCalled();
+        $adjustment2->setAdjustable($this)->shouldBeCalled();
+
+        $this->getAdjustmentsTotal()->shouldReturn(400);
+    }
+
+    function it_allows_to_know_amount_of_adjustments_by_type(
+        AdjustmentInterface $adjustment,
+        AdjustmentInterface $adjustment2
+    ) {
+        $adjustment->getAmount()->willReturn(100);
+        $adjustment2->getAmount()->willReturn(300);
+        $adjustment->getType()->willReturn('one');
+        $adjustment2->getType()->willReturn('two');
+        $adjustment->setAdjustable($this)->shouldBeCalled();
+        $adjustment2->setAdjustable($this)->shouldBeCalled();
+
+        $this->addAdjustment($adjustment);
+        $this->addAdjustment($adjustment2);
+
+        $this->getAdjustmentsTotal('one')->shouldReturn(100);
+        $this->getAdjustmentsTotal('two')->shouldReturn(300);
+    }
+
+    function it_allows_to_remove_adjustments_by_type(
+        AdjustmentInterface $adjustment,
+        AdjustmentInterface $adjustment2
+    ) {
+        $this->addAdjustment($adjustment);
+        $this->addAdjustment($adjustment2);
+        $adjustment->getType()->willReturn('one');
+        $adjustment2->getType()->willReturn('two');
+        $adjustment->isLocked()->willReturn(false);
+        $adjustment2->isLocked()->willReturn(false);
+
+        $this->getAdjustments()->shouldHaveCount(2);
+        $this->removeAdjustments('one');
+        $this->getAdjustments()->shouldHaveCount(1);
+        $this->removeAdjustments('two');
+        $this->getAdjustments()->shouldHaveCount(0);
+    }
+
+    function it_doesnt_remove_adjustment_if_these_are_locked(
+        AdjustmentInterface $adjustment,
+        AdjustmentInterface $adjustment2
+    ) {
+        $this->addAdjustment($adjustment);
+        $this->addAdjustment($adjustment2);
+
+        $adjustment->isLocked()->willReturn(false);
+        $adjustment2->isLocked()->willReturn(true);
+
+        $adjustment->getType()->willReturn('type');
+        $adjustment2->getType()->willReturn('type');
+
+        $this->getAdjustments()->shouldHaveCount(2);
+
+        $this->removeAdjustments('type');
+        $this->getAdjustments()->shouldHaveCount(1);
+    }
+
+    function it_allows_to_clear_its_adjustments(
+        AdjustmentInterface $adjustment,
+        AdjustmentInterface $adjustment2
+    ) {
+        $this->addAdjustment($adjustment);
+        $this->addAdjustment($adjustment2);
+
+        $this->clearAdjustments();
+
+        $this->getAdjustments()->shouldHaveCount(0);
     }
 }
