+++ src/Sylius/Component/Inventory/Model/InventoryUnitInterface.php
@@ -11,13 +11,12 @@
 
 namespace Sylius\Component\Inventory\Model;
 
-use Sylius\Component\Order\Model\AdjustableInterface;
 use Sylius\Component\Resource\Model\TimestampableInterface;
 
 /**
  * @author Paweł Jędrzejewski <pawel@sylius.org>
  */
-interface InventoryUnitInterface extends TimestampableInterface, AdjustableInterface
+interface InventoryUnitInterface extends TimestampableInterface
 {
     /**
      * Default states.
