+++ src/Sylius/Component/Order/Model/OrderInterface.php
@@ -94,9 +94,21 @@ interface OrderInterface extends
     public function getItemsTotal();
 
     /**
+     * Calculate items total based on the items
+     * unit prices and quantities.
+     */
+    public function calculateItemsTotal();
+
+    /**
      * @return integer
      */
     public function getTotal();
+
+    /**
+     * Calculate total.
+     * Items total + Adjustments total.
+     */
+    public function calculateTotal();
 
     /**
      * Alias of {@link countItems()}.
